mvs-designer/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
uuid==1.30
pathlib==1.0.1
subprocess32==3.5.4
psutil==5.9.6